Revert "[BREAKING CHANGE] Post Survey feature branch (#95)"
This reverts commit f2099a30ecd12d2d9b3db5ee7aa102edf9cfa1b1.
