+++ .gitignore
@@ -124,6 +124,5 @@ tech-matters-serverless-helpers/lib/
 
 
 .twilio-functions
-.twiliodeployinfo
 
 assets/blocklist.private.json
