--- assets/formDefinitions/v1/insights/postSurvey.private.json
@@ -1,12 +1,0 @@
-[
-  {
-    "insightsObject": "customers",
-    "attributeName": "customer_attribute_9",
-    "questions": ["wasHelpful"]
-  },
-  {
-    "insightsObject": "customers",
-    "attributeName": "customer_attribute_10",
-    "questions": ["wouldRecommend"]
-  }
-]
