+++ package.json
@@ -9,7 +9,7 @@
     "swagger": "redoc-cli bundle -o assets/swagger.html swagger.yaml",
     "lint": "eslint --ext ts .",
     "tr": "twilio-run",
-    "tr:deploy": "tsc && twilio-run deploy --force"
+    "tr:deploy": "tsc && npm run swagger && twilio-run deploy"
   },
   "repository": {
     "type": "git",
@@ -18,7 +18,6 @@
   "devDependencies": {
     "@types/jest": "^25.2.3",
     "@types/jsonwebtoken": "^8.5.0",
-    "@types/lodash": "^4.14.172",
     "@types/node-fetch": "^2.5.5",
     "@types/swagger-ui-express": "^4.1.1",
     "@types/twit": "^2.2.28",
@@ -37,7 +36,7 @@
     "prettier": "^1.19.1",
     "redoc-cli": "^0.10.2",
     "ts-jest": "^26.0.0",
-    "twilio-run": "^3.2.2",
+    "twilio-run": "2.* || >2.0.0-rc",
     "typescript": "^3.8.3"
   },
   "engines": {
@@ -46,12 +45,10 @@
   "dependencies": {
     "@tech-matters/serverless-helpers": "^1.0.6",
     "@twilio-labs/serverless-runtime-types": "^1.1.8",
-    "axios": "^0.21.1",
     "aws-sdk": "^2.975.0",
-    "lodash": "^4.17.21",
     "moment-timezone": "^0.5.33",
     "swagger-ui-express": "^4.1.4",
-    "twilio": "^3.67.0",
+    "twilio": "^3.60.0",
     "twilio-flex-token-validator": "^1.5.3",
     "twit": "^2.2.11",
     "type-fest": "^0.17.0"
